+++ switchbot_mqtt/_actors/__init__.py
@@ -19,8 +19,9 @@
 import logging
 import typing
 
-import bluepy.btle
 import aiomqtt
+import bleak
+import bleak.backends.device
 import switchbot
 
 from switchbot_mqtt._actors.base import _MQTTControlledActor
@@ -57,14 +58,16 @@ class _ButtonAutomator(_MQTTControlledActor):
     )
 
     def __init__(
-        self, *, mac_address: str, retry_count: int, password: typing.Optional[str]
+        self,
+        *,
+        device: bleak.backends.device.BLEDevice,
+        retry_count: int,
+        password: typing.Optional[str],
     ) -> None:
         self.__device = switchbot.Switchbot(
-            mac=mac_address, password=password, retry_count=retry_count
+            device=device, password=password, retry_count=retry_count
         )
-        super().__init__(
-            mac_address=mac_address, retry_count=retry_count, password=password
-        )
+        super().__init__(device=device, retry_count=retry_count, password=password)
 
     def _get_device(self) -> switchbot.SwitchbotDevice:
         return self.__device
@@ -79,7 +82,7 @@ class _ButtonAutomator(_MQTTControlledActor):
     ) -> None:
         # https://www.home-assistant.io/integrations/switch.mqtt/#payload_on
         if mqtt_message_payload.lower() == b"on":
-            if not self.__device.turn_on():
+            if not await self.__device.turn_on():
                 _LOGGER.error("failed to turn on switchbot %s", self._mac_address)
             else:
                 _LOGGER.info("switchbot %s turned on", self._mac_address)
@@ -95,7 +98,7 @@ class _ButtonAutomator(_MQTTControlledActor):
                     )
         # https://www.home-assistant.io/integrations/switch.mqtt/#payload_off
         elif mqtt_message_payload.lower() == b"off":
-            if not self.__device.turn_off():
+            if not await self.__device.turn_off():
                 _LOGGER.error("failed to turn off switchbot %s", self._mac_address)
             else:
                 _LOGGER.info("switchbot %s turned off", self._mac_address)
@@ -141,19 +144,21 @@ class _CurtainMotor(_MQTTControlledActor):
         )
 
     def __init__(
-        self, *, mac_address: str, retry_count: int, password: typing.Optional[str]
+        self,
+        *,
+        device: bleak.backends.device.BLEDevice,
+        retry_count: int,
+        password: typing.Optional[str],
     ) -> None:
         # > The position of the curtain is saved in self._pos with 0 = open and 100 = closed.
         # https://github.com/Danielhiversen/pySwitchbot/blob/0.10.0/switchbot/__init__.py#L150
         self.__device = switchbot.SwitchbotCurtain(
-            mac=mac_address,
+            device=device,
             password=password,
             retry_count=retry_count,
             reverse_mode=True,
         )
-        super().__init__(
-            mac_address=mac_address, retry_count=retry_count, password=password
-        )
+        super().__init__(device=device, retry_count=retry_count, password=password)
 
     def _get_device(self) -> switchbot.SwitchbotDevice:
         return self.__device
@@ -163,6 +168,7 @@ class _CurtainMotor(_MQTTControlledActor):
         mqtt_client: aiomqtt.Client,  # pylint: disable=duplicate-code; similar param list
         mqtt_topic_prefix: str,
     ) -> None:
+        assert self._basic_device_info is not None
         # > position_closed integer (Optional, default: 0)
         # > position_open integer (Optional, default: 100)
         # https://www.home-assistant.io/integrations/cover.mqtt/#position_closed
@@ -173,7 +179,7 @@ class _CurtainMotor(_MQTTControlledActor):
         await self._mqtt_publish(
             topic_prefix=mqtt_topic_prefix,
             topic_levels=self._MQTT_POSITION_TOPIC_LEVELS,
-            payload=str(int(self.__device.get_position())).encode(),
+            payload=str(int(self._basic_device_info["position"])).encode(),
             mqtt_client=mqtt_client,
         )
 
@@ -185,7 +191,7 @@ class _CurtainMotor(_MQTTControlledActor):
         report_position: bool = True,
     ) -> None:
         await super()._update_and_report_device_info(mqtt_client, mqtt_topic_prefix)
-        if report_position:
+        if self._basic_device_info and report_position:
             await self._report_position(
                 mqtt_client=mqtt_client, mqtt_topic_prefix=mqtt_topic_prefix
             )
@@ -201,7 +207,7 @@ class _CurtainMotor(_MQTTControlledActor):
         # https://www.home-assistant.io/integrations/cover.mqtt/#payload_open
         report_device_info, report_position = False, False
         if mqtt_message_payload.lower() == b"open":
-            if not self.__device.open():
+            if not await self.__device.open():
                 _LOGGER.error("failed to open switchbot curtain %s", self._mac_address)
             else:
                 _LOGGER.info("switchbot curtain %s opening", self._mac_address)
@@ -214,7 +220,7 @@ class _CurtainMotor(_MQTTControlledActor):
                 )
                 report_device_info = update_device_info
         elif mqtt_message_payload.lower() == b"close":
-            if not self.__device.close():
+            if not await self.__device.close():
                 _LOGGER.error("failed to close switchbot curtain %s", self._mac_address)
             else:
                 _LOGGER.info("switchbot curtain %s closing", self._mac_address)
@@ -226,7 +232,7 @@ class _CurtainMotor(_MQTTControlledActor):
                 )
                 report_device_info = update_device_info
         elif mqtt_message_payload.lower() == b"stop":
-            if not self.__device.stop():
+            if not await self.__device.stop():
                 _LOGGER.error("failed to stop switchbot curtain %s", self._mac_address)
             else:
                 _LOGGER.info("switchbot curtain %s stopped", self._mac_address)
@@ -269,7 +275,7 @@ class _CurtainMotor(_MQTTControlledActor):
         if message.retain:
             _LOGGER.info("ignoring retained message on topic %s", message.topic)
             return
-        actor = cls._init_from_topic(
+        actor = await cls._init_from_topic(
             topic=message.topic,
             mqtt_topic_prefix=mqtt_topic_prefix,
             expected_topic_levels=cls._MQTT_SET_POSITION_TOPIC_LEVELS,
@@ -284,7 +290,7 @@ class _CurtainMotor(_MQTTControlledActor):
             _LOGGER.warning("invalid position %u%%, ignoring message", position_percent)
             return
         # pylint: disable=protected-access; own instance
-        if actor._get_device().set_position(position_percent):
+        if await actor._get_device().set_position(position_percent):
             _LOGGER.info(
                 "set position of switchbot curtain %s to %u%%",
                 actor._mac_address,